moved writedb to common
